Fix discard failure
